Unbundeled SYCL from main library
SYCL is now built as a shared library. This means that if one doesn't
use the `sycl` or `gpu` backend missing runtime libraries (e.g. CUDA or
HIP) won't affect the `cpu` backend. This allows one to build the
library with SYCL without needing the runtime libraries to be present
for every backend.